refactor init_module function
